ci(generate): do not run py38-path tox env & improve ci steps names
Previously, the generate-workflow included the py38-path tox env as part of the 'test against
various installations' step. Now, the step mentioned does not include the py38-path tox env. This
currently helps prevents generating a poetry.lock file which the following steps accidentally commit
and push to the remote 'origin' of the biskotaki repository.