feat(py<3.10): lift requirement of runtime py version >=3.10, for 'interpreters interactive dialog'
Previously, at runtime, the 'interactive-cli-interpreters-dialog' feature would only be available,
if running on python >=3.10 (assuming running in 'interactive mode' with `--no-input=False`). For
python 3.8, and 3.9 the feature was disabled. So, prevously if CLi was running on python 3.8, or 3.9
and 'interactive mode' was enabled, then Generator fallback back, for
'interactive-cli-interpreters-dialog', would be the dialog that Cookiecutter uses to prompt for
input. Now, since we have already migrated from the problematic PyInquirer library to the
Questionary, our Interactive Interpreters Dialog implementation is not restricted for python <3.10,
and we can advertize that Interactive Interpreters Dialog is now supported accross all Python
versions that our package allows, through pyproject.toml. At the moment they are ">3.6,<3.13".