test: Regression and Backwards Compatibility
Add Automated Regression Test for 'Generator Docs Default Settings'. By Default Generator v2.x promises that upon invocation, the generated Project with be configured for Doc Builds with Sphinx and RTD Python3.8. The Test will fail if Defaults change. If Test is fixed, then this is evidence that if changes are released, they possibly introduce "breaking changes". Also, we add a new Test on the Dialog Prompts to verify that they prompt User for each and every one of the Cookiecutter Tempalte Variables. If this test fails, then the Dialog System and the Cookiecutter Tempalte Vars are out of Sync, which should be addressed, preventing a Bug from shipping Code with the 2 components being NOT synced. From this commit and on, we also stop include Code Coverage for Test Code ('tests' dir), but rather only for the Production Code, in 'src' dir.
Tests, for automatic compati and increase  Testsmeasure Code Coverage only for Code in 'src', and not 'tests', dir

test(dialog-system): verify prompts user for each and everyone of Cookiecutter Template Variables

test(docs-settings): verify Gen Project configured with Sphinx and RTD Py3.8, as advertised Defaults

refactor(isort): apply isort

refactor(black): apply black